apps/api/pyproject.toml
[project]
name = "byteboost-api"
version = "0.1.0"
description = "ByteBoost E-Learning Platform API"
authors = [{name = "ByteBoost Team", email = "team@byteboost.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy>=2.0.30",
    "psycopg[binary]>=3.1.19",
    "alembic>=1.13.0",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "redis>=5.0.0",
    "celery>=5.4.0",
    "boto3>=1.34.0",
    "razorpay>=1.4.1",
    "slowapi>=0.1.9",
    "sentry-sdk[fastapi]>=2.0.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "black>=24.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "faker>=25.0.0"
]

[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["E501", "B008", "C901", "W191"]
exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    ".mypy_cache",
    ".pytest_cache"
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true